server has been modified

+++ routes/shopping-carts.js
@@ -3,10 +3,14 @@ const {
   ShoppingCart,
   validateItem,
   validateShoppingCart,
+  removeItem,
+  updateItem,
+  createItem,
+  calculateTotalPrice,
+  addItemsToArray
 } = require('../models/shopping-cart');
 const { Product } = require('../models/product');
 const { User } = require('../models/user');
-const { Order } = require('../models/order');
 const express = require('express');
 const router = express.Router();
 
@@ -21,7 +25,10 @@ router.get('/', async (req, res) => {
 
 router.get('/:id', async (req, res) => {
   try {
-    //const shoppingCarts = await ShoppingCart.find({ 'customer._id': req.body.userId });
+    if (!mongoose.Types.ObjectId.isValid(req.params.id))
+      return send.status(400)
+        .send('Invalid ShoppingCart.');
+
     let shoppingCart = await ShoppingCart.findById(req.params.id);
     if (!shoppingCart)
       return res
@@ -42,17 +49,17 @@ router.post('/', async (req, res) => {
     const user = await User.findById(req.body.userId);
     if (!user) return res.status(400).send('Invalid User.');
 
-    //uncomment after creation any order
-    //const orders = await Order.findById(req.body.userId);
-    const orders = [];
+
 
     const shoppingCarts = await ShoppingCart.find({
-      'customer._id': req.body.userId,
+      $and: [
+        { 'customer._id': req.body.userId },
+        { orderId: { $eq: null } }
+      ]
     });
 
-    //check if there is shoppingCart without order 
-    //return with status 400 the shoppingCart._id
-    if (shoppingCarts.length > orders.length)
+
+    if (shoppingCarts.length > 0)
       return res.status(400).send('The user has already the shopping cart');
 
     let itemsArr = [];
@@ -65,37 +72,13 @@ router.post('/', async (req, res) => {
           .status(400)
           .send('The product with given ID has not been found');
 
-          itemsArr = addItemsToArray();
+      itemsArr = addItemsToArray(products, req.body.items);
 
-      // products.forEach((prod) => {
-      //   const productInCart = req.body.items.filter(
-      //     (e) => e.productId === prod._id.toString()
-      //   );
-
-      //   let totalPrice = prod.price * productInCart[0].quantity;
-      //   item = {
-      //     product: {
-      //       _id: prod._id,
-      //       price: prod.price,
-      //       title: prod.title,
-      //       category: prod.category,
-      //       imageUrl: prod.imageUrl,
-      //     },
-      //     quantity: productInCart[0].quantity,
-      //     itemTotalPrice: totalPrice,
-      //   };
-
-      //   itemsArr.push(item);
-      // });
     }
 
     let total = 0;
-
     if (itemsArr.length > 0)
-      total = itemsArr.reduce(
-        (accumulator, item) => accumulator + item.itemTotalPrice,
-        total
-      );
+     total = calculateTotalPrice(itemsArr);
 
     let shoppingCart = new ShoppingCart({
       customer: {
@@ -106,7 +89,9 @@ router.post('/', async (req, res) => {
       items: itemsArr,
       totalPrice: total,
     });
+
     shoppingCart = await shoppingCart.save();
+
     res.send(shoppingCart);
   } catch (e) {
     throw new Error(e);
@@ -115,36 +100,16 @@ router.post('/', async (req, res) => {
 });
 
 
-function addItemsToArray(products) {
-  let itemsArr = [];
-  products.forEach((prod) => {
-    const productInCart = req.body.items.filter(
-      (e) => e.productId === prod._id.toString()
-    );
-
-    let totalPrice = prod.price * productInCart[0].quantity;
-    item = {
-      product: {
-        _id: prod._id,
-        price: prod.price,
-        title: prod.title,
-        category: prod.category,
-        imageUrl: prod.imageUrl,
-      },
-      quantity: productInCart[0].quantity,
-      itemTotalPrice: totalPrice,
-    };
-
-    itemsArr.push(item);
-  });
-  return itemsArr;
-}
-
 router.patch('/:id', async (req, res) => {
   const { error } = validateItem(req.body);
   if (error) res.status(400).send(error.details[0].message);
 
   try {
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id))
+      return send.status(400)
+        .send('Invalid ShoppingCart.');
+
     let shoppingCart = await ShoppingCart.findById(req.params.id);
 
     if (!shoppingCart)
@@ -163,42 +128,19 @@ router.patch('/:id', async (req, res) => {
       (item) => item.product._id.toString() === product._id.toString()
     );
 
-    let item;
     const productQuantity = +req.body.quantity;
     if (!productInCart) {
-      let productTotalPrice = productQuantity * product.price;
-      item = {
-        product: {
-          _id: product._id,
-          price: product.price,
-          title: product.title,
-          category: product.category,
-          imageUrl: product.imageUrl,
-        },
-        quantity: productQuantity,
-        itemTotalPrice: productTotalPrice,
-      };
+      const item = createItem(product, productQuantity);
       shoppingCart.items.push(item);
     } else {
-      if (productQuantity === 0) {
-        shoppingCart.items.id(shoppingCart.items[indexOfProduct]._id).remove();
-      } else {
-        shoppingCart.items[indexOfProduct].product.price = product.price;
-        shoppingCart.items[indexOfProduct].quantity = productQuantity;
-        shoppingCart.items[indexOfProduct].itemTotalPrice =
-          productQuantity * product.price;
-      }
+      if (productQuantity === 0)
+        removeItem(shoppingCart, indexOfProduct);
+      else
+        updateItem(shoppingCart, product.price, productQuantity, indexOfProduct);
     }
 
-    let total = 0;
+    shoppingCart.totalPrice = calculateTotalPrice(shoppingCart.items);
 
-    if (shoppingCart.items.length > 0)
-      total = shoppingCart.items.reduce(
-        (accumulator, item) => accumulator + +item.itemTotalPrice,
-        total
-      );
-
-    shoppingCart.totalPrice = total;
     shoppingCart = await shoppingCart.save();
 
     res.send(shoppingCart);
@@ -210,6 +152,11 @@ router.patch('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
   try {
+
+    if (!mongoose.Types.ObjectId.isValid(req.params.id))
+      return send.status(400)
+        .send('Invalid ShoppingCart.');
+
     let shoppingCart = await ShoppingCart.findByIdAndRemove(req.params.id);
     if (!shoppingCart)
       res.status(404).send('The shoppingCart with given ID has not been found');
@@ -221,3 +168,6 @@ router.delete('/:id', async (req, res) => {
 });
 
 module.exports = router;
+
+
+
